refactor: centralize network configuration and skip non-Windows tests
