+++ DesktopApplicationTemplate.Core/Services/IProcessRunner.cs
@@ -1,7 +1,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace DesktopApplicationTemplate.UI.Services
+namespace DesktopApplicationTemplate.Core.Services
 {
     public interface IProcessRunner
     {
