+++ DesktopApplicationTemplate.Core/Services/ProcessRunner.cs
@@ -3,7 +3,7 @@ using System.Diagnostics;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace DesktopApplicationTemplate.UI.Services
+namespace DesktopApplicationTemplate.Core.Services
 {
     public class ProcessRunner : IProcessRunner
     {
