+++ DesktopApplicationTemplate.Tests/LogLevelSelectionTests.cs
@@ -18,8 +18,7 @@ namespace DesktopApplicationTemplate.Tests
         [WpfFact]
         public void HttpServiceView_LogLevelSelection_UpdatesLogger()
         {
-            if (!OperatingSystem.IsWindows())
-                return;
+            Skip.IfNot(OperatingSystem.IsWindows(), "Requires Windows desktop runtime");
 
             Exception? ex = null;
             var thread = new Thread(() =>
