+++ DesktopApplicationTemplate.Tests/MainViewCreateNavigationTests.cs
@@ -1,10 +1,11 @@
 using System;
 using System.IO;
 using System.Threading;
+using DesktopApplicationTemplate.UI.Services;
 using DesktopApplicationTemplate.Core.Services;
 using DesktopApplicationTemplate.Service.Services;
-using DesktopApplicationTemplate.UI.Models;
-using DesktopApplicationTemplate.UI.Services;
+using DesktopApplicationTemplate.Core.Models;
+using DesktopApplicationTemplate.Core.Services;
 using DesktopApplicationTemplate.UI.ViewModels;
 using DesktopApplicationTemplate.UI.Views;
 using DesktopApplicationTemplate.UI;
@@ -22,8 +23,7 @@ public class MainViewCreateNavigationTests
     [WpfFact]
     public void NavigateToTcp_ShowsCreateView()
     {
-        if (!OperatingSystem.IsWindows())
-            return;
+        Skip.IfNot(OperatingSystem.IsWindows(), "Requires Windows desktop runtime");
 
         var thread = new Thread(() =>
         {
@@ -67,8 +67,7 @@ public class MainViewCreateNavigationTests
     [WpfFact]
     public void NavigateToFtpServer_ShowsCreateView()
     {
-        if (!OperatingSystem.IsWindows())
-            return;
+        Skip.IfNot(OperatingSystem.IsWindows(), "Requires Windows desktop runtime");
 
         var thread = new Thread(() =>
         {
@@ -114,8 +113,7 @@ public class MainViewCreateNavigationTests
     [WpfFact]
     public void NavigateToHid_ShowsCreateView()
     {
-        if (!OperatingSystem.IsWindows())
-            return;
+        Skip.IfNot(OperatingSystem.IsWindows(), "Requires Windows desktop runtime");
 
         var thread = new Thread(() =>
         {
@@ -163,8 +161,7 @@ public class MainViewCreateNavigationTests
     [WpfFact]
     public void NavigateToFileObserver_ShowsCreateView()
     {
-        if (!OperatingSystem.IsWindows())
-            return;
+        Skip.IfNot(OperatingSystem.IsWindows(), "Requires Windows desktop runtime");
 
         var thread = new Thread(() =>
         {
@@ -213,8 +210,7 @@ public class MainViewCreateNavigationTests
     [WpfFact]
     public void NavigateToScp_ShowsCreateView()
     {
-        if (!OperatingSystem.IsWindows())
-            return;
+        Skip.IfNot(OperatingSystem.IsWindows(), "Requires Windows desktop runtime");
 
         var thread = new Thread(() =>
         {
