+++ DesktopApplicationTemplate.Tests/MainViewHttpNavigationTests.cs
@@ -1,9 +1,10 @@
 using System;
 using System.IO;
 using System.Threading;
-using DesktopApplicationTemplate.Core.Services;
-using DesktopApplicationTemplate.UI.Models;
 using DesktopApplicationTemplate.UI.Services;
+using DesktopApplicationTemplate.Core.Services;
+using DesktopApplicationTemplate.Core.Models;
+using DesktopApplicationTemplate.Core.Services;
 using DesktopApplicationTemplate.UI.ViewModels;
 using DesktopApplicationTemplate.UI.Views;
 using DesktopApplicationTemplate.UI;
@@ -22,8 +23,7 @@ public class MainViewHttpNavigationTests
     [WpfFact]
     public void NavigateToHttp_ShowsCreateView()
     {
-        if (!OperatingSystem.IsWindows())
-            return;
+        Skip.IfNot(OperatingSystem.IsWindows(), "Requires Windows desktop runtime");
 
         var thread = new Thread(() =>
         {
@@ -68,8 +68,7 @@ public class MainViewHttpNavigationTests
     [WpfFact]
     public void EditHttpService_ShowsEditView()
     {
-        if (!OperatingSystem.IsWindows())
-            return;
+        Skip.IfNot(OperatingSystem.IsWindows(), "Requires Windows desktop runtime");
 
         var thread = new Thread(() =>
         {
@@ -123,8 +122,7 @@ public class MainViewHttpNavigationTests
     [WpfFact]
     public void DoubleClickHttpService_OpensEditView()
     {
-        if (!OperatingSystem.IsWindows())
-            return;
+        Skip.IfNot(OperatingSystem.IsWindows(), "Requires Windows desktop runtime");
 
         var thread = new Thread(() =>
         {
