+++ DesktopApplicationTemplate.Tests/MainViewTcpEditTests.cs
@@ -2,11 +2,10 @@ using System;
 using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
-using DesktopApplicationTemplate.Core.Services;
 using DesktopApplicationTemplate.UI.Services;
-using DesktopApplicationTemplate.UI.ViewModels;
+using DesktopApplicationTemplate.Core.Services;
 using DesktopApplicationTemplate.UI.Views;
-using DesktopApplicationTemplate.UI.Models;
+using DesktopApplicationTemplate.Core.Models;
 using DesktopApplicationTemplate.UI;
 using DesktopApplicationTemplate.UI.Helpers;
 using FluentAssertions;
