+++ DesktopApplicationTemplate.Tests/NetworkConfigurationViewModelTests.cs
@@ -1,6 +1,6 @@
 using System.Threading.Tasks;
-using DesktopApplicationTemplate.UI.Models;
-using DesktopApplicationTemplate.UI.Services;
+using DesktopApplicationTemplate.Core.Models;
+using DesktopApplicationTemplate.Core.Services;
 using DesktopApplicationTemplate.UI.ViewModels;
 using Moq;
 using Xunit;
