+++ DesktopApplicationTemplate.Tests/ThemeManagerTests.cs
@@ -10,10 +10,7 @@ namespace DesktopApplicationTemplate.Tests
         [WpfFact]
         public void ApplyTheme_LoadsResourceDictionary()
         {
-            if (!OperatingSystem.IsWindows())
-            {
-                return;
-            }
+            Skip.IfNot(OperatingSystem.IsWindows(), "Requires Windows desktop runtime");
 
             Exception? ex = null;
             var thread = new Thread(() =>
